feat(assertion): implemented the core assertion functions
The validator architecture is unfortunately not flexible. For example,
let's say we need to check if a solution spawned a process. We would
have to write a new validator and a corresponding GUI. If we need to
check the contents of a file, we need to create another validator
and a corresponding GUI. As you can see, for each new requirement,
we keep creating new validators.

Now let's push the test cases further. We need to check for a condition
before and after the solution is run. Such edge cases will require us
to patch our validators over and over.

In light of this realization, we are letting go of this architecture.
Now all the test cases are scripts that are bundled with Rover.
If AcademyJS has 1000 excercises, Rover will come with 1000 test cases.
This is not really a problem because we will hit a saturation point
eventually.

 * Removed the `solving` command. The user can just use the `submit`
   command for both identifying and submiting a solution.
 * The standard output and error data are now neatly formatted.

BREAKING CHANGES: `--std-error` and `--std-output` options are ignored.